refactor: change checkUserData helper
