--- src/helpers/getUserIdFromLocalStorage.ts
@@ -1,3 +1,0 @@
-export function getUserIdFromLocalStorage(): string | false {
-  return localStorage.getItem('userId') || false;
-}
