+++ src/helpers/index.ts
@@ -1,1 +1,2 @@
 export { getEnvValue } from './getEnvValue';
+export { checkUserIdandRoomId } from './checkUserData';
