--- src/helpers/setUserIdToLocalStorage.ts
@@ -1,3 +1,0 @@
-export function setUserIdToLocalStorage(id: string): void {
-  localStorage.setItem('userId', id);
-}
